+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 1.5.0 (2025-04-04)
 ------------------
 
+- #125 Add Age field for patient content
 - #123 Hyperlink patient on sample listing
 - #119 Rely on `catalog_mappings` to get the catalogs for Patient portal type
 - #117 Make accessor and mutator from Patient to rely on base class